+++ apps/content-publishing-worker/src/batch_announcer/batch.announcer.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { BatchAnnouncementService } from './batch.announcer.service';
 import { BatchAnnouncer } from './batch.announcer';
-import { BlockchainModule } from '#content-publishing-lib/blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 
 @Module({
