+++ apps/content-publishing-worker/src/batching_processor/batching.processor.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ScheduleModule } from '@nestjs/schedule';
 import { BatchingProcessorService } from './batching.processor.service';
 import { BroadcastWorker, ReplyWorker, ReactionWorker, TombstoneWorker, UpdateWorker, ProfileWorker } from './workers';
-import { BlockchainModule } from '#content-publishing-lib/blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 
 @Module({
   imports: [ScheduleModule.forRoot(), BlockchainModule],
