+++ apps/content-publishing-worker/src/publisher/publisher.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PublishingService } from './publishing.service';
 import { IPFSPublisher } from './ipfs.publisher';
-import { NonceService } from './nonce.service';
-import { BlockchainModule } from '#content-publishing-lib/blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 
 @Module({
   imports: [BlockchainModule],
   controllers: [],
-  providers: [PublishingService, IPFSPublisher, NonceService],
+  providers: [PublishingService, IPFSPublisher],
   exports: [PublishingService, IPFSPublisher],
 })
 export class PublisherModule {}
