--- libs/account-lib/src/interfaces/capacity-limit.interface.ts
@@ -1,9 +1,0 @@
-export interface ICapacityLimit {
-  type: 'percentage' | 'amount';
-  value: bigint;
-}
-
-export interface ICapacityLimits {
-  serviceLimit: ICapacityLimit;
-  totalLimit?: ICapacityLimit;
-}
