--- libs/content-publishing-lib/src/blockchain/blockchain.config.spec.ts
@@ -1,88 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { describe, it, expect, beforeAll } from '@jest/globals';
-import blockchainConfig, { allowReadOnly, IBlockchainConfig } from './blockchain.config';
-import configSetup from '#testlib/utils.config-tests';
-
-const { setupConfigService, validateMissing, shouldFailBadValues } = configSetup<IBlockchainConfig>(blockchainConfig);
-const { setupConfigService: setupConfigServiceReadOnly } = configSetup<IBlockchainConfig>(allowReadOnly);
-
-describe('Blockchain module config', () => {
-  const ALL_ENV: { [key: string]: string | undefined } = {
-    FREQUENCY_URL: undefined,
-    PROVIDER_ACCOUNT_SEED_PHRASE: undefined,
-    PROVIDER_ID: undefined,
-    CAPACITY_LIMIT: undefined,
-  };
-
-  beforeAll(() => {
-    Object.keys(ALL_ENV).forEach((key) => {
-      ALL_ENV[key] = process.env[key];
-    });
-  });
-
-  describe('invalid environment', () => {
-    it('missing frequency url should fail', async () => validateMissing(ALL_ENV, 'FREQUENCY_URL'));
-    it('invalid frequency url should fail', async () => shouldFailBadValues(ALL_ENV, 'FREQUENCY_URL', ['invalid url']));
-
-    it('missing capacity limits should fail', async () => validateMissing(ALL_ENV, 'CAPACITY_LIMIT'));
-
-    it('invalid capacity limit should fail', async () =>
-      shouldFailBadValues(ALL_ENV, 'CAPACITY_LIMIT', [
-        '{ "type": "bad type", "value": 0 }',
-        '{ "type": "percentage", "value": -1 }',
-        '{ "type": "percentage", "value": 101 }',
-        '{ "type": "amount", "value": -1 }',
-      ]));
-
-    it('missing provider ID should fail', async () => validateMissing(ALL_ENV, 'PROVIDER_ID'));
-
-    it('missing seed phrase should fail', async () => validateMissing(ALL_ENV, 'PROVIDER_ACCOUNT_SEED_PHRASE'));
-
-    it('missing seed phrase should activate read-only mode if allowed', async () => {
-      const { PROVIDER_ACCOUNT_SEED_PHRASE: _dummy, ...env } = ALL_ENV;
-      const accountConfigService = await setupConfigServiceReadOnly(env);
-      expect(accountConfigService.isDeployedReadOnly).toBeTruthy();
-    });
-
-    it('empty seed phrase should activate read-only mode if allowed', async () => {
-      const { PROVIDER_ACCOUNT_SEED_PHRASE: _dummy, ...env } = ALL_ENV;
-      const accountConfigService = await setupConfigServiceReadOnly({ PROVIDER_ACCOUNT_SEED_PHRASE: '', ...env });
-      expect(accountConfigService.isDeployedReadOnly).toBeTruthy();
-    });
-
-    it('blank seed phrase should activate read-only mode if allowed', async () => {
-      const { PROVIDER_ACCOUNT_SEED_PHRASE: _dummy, ...env } = ALL_ENV;
-      const accountConfigService = await setupConfigServiceReadOnly({ PROVIDER_ACCOUNT_SEED_PHRASE: '    ', ...env });
-      expect(accountConfigService.isDeployedReadOnly).toBeTruthy();
-    });
-  });
-
-  describe('valid environment', () => {
-    let blockchainConf: IBlockchainConfig;
-    beforeAll(async () => {
-      blockchainConf = await setupConfigService(ALL_ENV);
-    });
-
-    it('should be defined', () => {
-      expect(blockchainConf).toBeDefined();
-    });
-
-    it('should get frequency url', () => {
-      const expectedUrl = new URL(ALL_ENV.FREQUENCY_URL).toString();
-      expect(blockchainConf.frequencyUrl?.toString()).toStrictEqual(expectedUrl);
-    });
-
-    it('should get provider account seed phrase and not be readonly', () => {
-      expect(blockchainConf.providerSeedPhrase).toStrictEqual(ALL_ENV.PROVIDER_ACCOUNT_SEED_PHRASE);
-      expect(blockchainConf.isDeployedReadOnly).toStrictEqual(false);
-    });
-
-    it('should get provider id', () => {
-      expect(blockchainConf.providerId.toString()).toStrictEqual(ALL_ENV.PROVIDER_ID);
-    });
-
-    it('should get capacity limit', () => {
-      expect(JSON.stringify(blockchainConf.capacityLimit)).toStrictEqual(ALL_ENV.CAPACITY_LIMIT!);
-    });
-  });
-});
