+++ package.json
@@ -90,12 +90,12 @@
     "test:k6:account": "for t in apps/account-api/k6-test/*.k6.js ; do k6 run ${t} ; done",
     "format": "prettier --write .",
     "lint": "eslint apps/**/*.[tj]s libs/**/*.[tj]s",
-    "lint:account": "eslint apps/account*/**/*.[tj]s libs/account-lib/**/*.[tj]s",
-    "lint:content-publishing": "eslint apps/content-publishing*/**/*.[tj]s libs/content-publishing-lib/**/*.[tj]s",
-    "lint:content-watcher": "eslint apps/content-watcher/**/*.[tj]s libs/content-watcher-lib/**/*.[tj]s",
-    "lint:graph": "eslint apps/graph*/**/*.[tj]s libs/graph-lib/**/*.[tj]s",
-    "lint:libs": "eslint libs/**/*.[tj]s",
-    "lint:fix": "eslint --fix apps/**/*.[tj]s libs/**/*.[tj]s",
+    "lint:account": "eslint 'apps/account*/**/*.[tj]s' 'libs/account-lib/**/*.[tj]s'",
+    "lint:content-publishing": "eslint 'apps/content-publishing*/**/*.[tj]s' 'libs/content-publishing-lib/**/*.[tj]s'",
+    "lint:content-watcher": "eslint 'apps/content-watcher/**/*.[tj]s' 'libs/content-watcher-lib/**/*.[tj]s'",
+    "lint:graph": "eslint 'apps/graph*/**/*.[tj]s' 'libs/graph-lib/**/*.[tj]s'",
+    "lint:libs": "eslint 'libs/**/*.[tj]s'",
+    "lint:fix": "eslint --fix 'apps/**/*.[tj]s' 'libs/**/*.[tj]s'",
     "setup:content-publishing:chain": "node scripts/content-publishing/local-init.cjs",
     "setup:content-watcher:chain": "node scripts/content-watcher/local-chain-setup.cjs",
     "setup:content-watcher:publish": "node scripts/content-watcher/content-setup.js",
